Migrate logger to caddy2
