+++ go/cmd/pfhttpd2/main.go
@@ -10,6 +10,7 @@ import (
 	_ "github.com/inverse-inc/packetfence/go/plugin/caddy2/httpdispatcher"
 	_ "github.com/inverse-inc/packetfence/go/plugin/caddy2/httpdportalpreview"
 	_ "github.com/inverse-inc/packetfence/go/plugin/caddy2/job-status"
+	_ "github.com/inverse-inc/packetfence/go/plugin/caddy2/logger"
 	_ "github.com/inverse-inc/packetfence/go/plugin/caddy2/pfconfig"
 )
